# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Fuyao Procurement Cloud API
  version: 0.1.0
  description: >
    Backend contract for the Fuyao procurement SaaS. This mirrors exactly what
    the frontend calls in `src/lib/api.ts`; the data shapes come from
    `src/types.ts` (the single source of truth). Implement this and the
    frontend works with no changes — just set `VITE_API_BASE_URL` to your URL.


    Conventions:
      * REST + JSON, all paths under `/api`.
      * Auth: every endpoint except `POST /api/auth/login` requires
        `Authorization: Bearer <JWT>`.
      * Errors use the envelope `{ "error": { "code", "message" } }`.
      * Timestamps are epoch MILLISECONDS (numbers), matching JS `Date.now()`.
        If you store `timestamptz` in Postgres, convert to ms on the way out.
      * `id` fields are strings (UUID recommended).
servers:
  - url: '{baseUrl}'
    description: Backend base URL (set as VITE_API_BASE_URL in the frontend)
    variables:
      baseUrl:
        default: http://localhost:8000
security:
  - bearerAuth: []
...
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Sign in, returns a JWT and the user profile
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Error'
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Current user for the bearer token (used to restore a session)
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUser'
        '401':
          $ref: '#/components/responses/Error'
...
  /api/sourcing/search:
    post:
      tags: [Sourcing]
      summary: Find suppliers for a natural-language requirement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourcingSearchRequest'
      responses:
        '200':
          description: Matching suppliers
          content:
            application/json:
              schema:
                type: object
                required: [results]
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/Supplier'
  /api/sourcing/search-jobs:
    post:
      tags: [Sourcing]
      summary: Start an asynchronous supplier research job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourcingSearchRequest'
      responses:
        '202':
          description: Search job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchJobResponse'
  /api/sourcing/search-jobs/{job_id}:
    get:
      tags: [Sourcing]
      summary: Poll an asynchronous supplier research job
      parameters:
        - name: job_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Current job state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchJobResponse'
        '404':
          $ref: '#/components/responses/Error'
  /api/sourcing/search-jobs/{job_id}/events:
    get:
      tags: [Sourcing]
      summary: Stream supplier research job progress as Server-Sent Events
      parameters:
        - name: job_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Server-Sent Events containing SearchJobResponse payloads
          content:
            text/event-stream:
              schema:
                type: string
        '404':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: INVALID_CREDENTIALS
            message:
              type: string
              example: Email or password is incorrect
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
          description: 'JWT to send as Authorization: Bearer <token>'
        user:
          $ref: '#/components/schemas/AuthUser'
    AuthUser:
      type: object
      required: [email, name, company, role]
      properties:
        email: { type: string }
        name: { type: string }
        company: { type: string }
        role: { type: string }
...
    SourcingSearchRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
          example: Water deflector strips and glass adhesive in Germany
        structured:
          $ref: '#/components/schemas/SourcingStructuredFields'
        criteria:
          type: array
          maxItems: 12
          description: User-configurable supplier evaluation dimensions and weights.
          items: { $ref: '#/components/schemas/EvaluationCriterion' }
    SourcingStructuredFields:
      type: object
      description: Optional structured sourcing fields; they override/augment parsed intent.
      properties:
        productName: { type: string }
        quantity: { type: string }
        unit: { type: string }
        brand: { type: string }
        model: { type: string }
        specifications: { type: string }
        standards: { type: string, description: 'Comma-separated standards / certifications' }
        category: { type: string }
        country: { type: string }
        targetRegion: { type: string }
        certifications: { type: string }
        minOrderQuantity: { type: string }
        qualityRequirements: { type: string }
        environmentalRequirements: { type: string }
    EvaluationCriterion:
      type: object
      required: [key, weight]
      properties:
        key:
          type: string
          description: Stable machine key, e.g. quality, reputation, minimumOrderQuantity, environmental.
          pattern: '^[A-Za-z][A-Za-z0-9_-]*$'
        label: { type: string }
        weight: { type: number, minimum: 0, maximum: 100 }
    SupplierCategory:
      type: string
      description: Open category string. Known examples include waterDeflector, glassAdhesive, rubberSeal, glassRaw, hardware, packaging, cleaning, office, safetyShoes, firstAid, equipment.
    Supplier:
      type: object
      required:
        [id, name, category, country, city, address, contactPerson, phone, email,
         website, employees, annualRevenue, established, capabilities, certifications, matchScore]
      properties:
        id: { type: string }
        name: { type: string }
        category: { $ref: '#/components/schemas/SupplierCategory' }
        country: { type: string }
        city: { type: string }
        address: { type: string }
        contactPerson: { type: string }
        phone: { type: string }
        email: { type: string }
        website: { type: string }
        employees: { type: string, description: 'Headcount band, e.g. "250–500"' }
        annualRevenue: { type: string, description: 'e.g. "€ 80M–120M"' }
        established: { type: integer, example: 1995 }
        capabilities:
          type: array
          items: { type: string }
        certifications:
          type: array
          items: { type: string }
        productName: { type: string, nullable: true }
        brand: { type: string, nullable: true }
        model: { type: string, nullable: true }
        specifications: { type: string, nullable: true }
        standards:
          type: array
          items: { type: string }
        unitPriceEur: { type: number, nullable: true }
        unitPrice: { type: number, nullable: true, description: 'Compatibility alias for unitPriceEur' }
        currency: { type: string, example: EUR }
        unitLabel: { type: string }
        quoteConditions: { type: string, nullable: true }
        deliveryDays: { type: integer, nullable: true }
        deliveryLabel: { type: string }
        deliveryLeadTime: { type: string, description: 'Compatibility alias for deliveryLabel' }
        paymentTerm: { type: string, nullable: true }
        paymentLabel: { type: string }
        paymentTerms: { type: string, description: 'Compatibility alias for paymentLabel' }
        source: { type: string, enum: [database, web] }
        sourceDetail: { type: string }
        sourceUrls:
          type: array
          items: { type: string }
        evidenceSnippets:
          type: array
          items: { type: string }
        verificationStatus: { type: string, enum: [verified, needs-review] }
        verificationNotes:
          type: array
          items: { type: string }
        dataCompleteness: { type: number, minimum: 0, maximum: 100 }
        criteriaScores:
          type: object
          additionalProperties: { type: number, minimum: 0, maximum: 100 }
        weightedCriteriaScore: { type: number, minimum: 0, maximum: 100 }
        matchScore:
          type: number
          minimum: 0
          maximum: 100
          description: Relevance to the query (0–100)
...
    SearchJobEvent:
      type: object
      required: [timestamp, phase, message, progress]
      properties:
        timestamp: { type: integer, format: int64, description: Epoch milliseconds }
        phase: { type: string }
        message: { type: string }
        progress: { type: integer, minimum: 0, maximum: 100 }
    SearchJobResponse:
      type: object
      required: [jobId, status, progress, step, events, results]
      properties:
        jobId: { type: string }
        status:
          type: string
          enum: [queued, running, completed, failed]
        progress: { type: integer, minimum: 0, maximum: 100 }
        step: { type: string }
        events:
          type: array
          items: { $ref: '#/components/schemas/SearchJobEvent' }
        intent:
          type: object
          nullable: true
          additionalProperties: true
        results:
          type: array
          items: { $ref: '#/components/schemas/Supplier' }
        error:
          type: string
          nullable: true
...