# src/types.ts
// ---------------------------------------------------------------------------
// Shared domain types for the Fuyao procurement SaaS shell.
// NOTE: This is a front-end-only prototype. Everything below is wired to mock
// JSON data; the shapes are designed so a real backend can drop in later.
// ---------------------------------------------------------------------------

export type Language = 'en' | 'zh'

export type ModuleId = 'sourcing' | 'comparison' | 'memory' | 'settings'

// --- Auth ------------------------------------------------------------------

export interface AuthUser {
  email: string
  name: string
  company: string
  role: string
}

/** A secret stored in the (mock) cloud key vault. Value is never shown raw. */
export interface VaultKey {
  id: string
  label: string
  /** Last 4 chars only, for display. The full value is treated as write-only. */
  maskedValue: string
  updatedAt: number
}

// --- Sourcing (supplier discovery) -----------------------------------------

export type SupplierCategoryKey =
  | 'waterDeflector' // 挡水条
  | 'glassAdhesive' // 玻璃胶 / urethane
  | 'rubberSeal' // 密封条 / weatherstrip
  | 'glassRaw' // 玻璃原片 / float glass
  | 'hardware' // 五金件 / mounting hardware
  | 'packaging' // 包装材料
  | 'cleaning'
  | 'office'
  | 'safetyShoes'
  | 'firstAid'
  | 'equipment'
  | string

export interface Supplier {
  id: string
  name: string
  category: SupplierCategoryKey
  /** Detail fields below are optional: real (scraped) data may not have them. */
  country?: string
  city?: string
  address?: string
  contactPerson?: string
  phone?: string
  email?: string
  website?: string
  /** Headcount band, e.g. "250–500". */
  employees?: string
  /** Annual revenue band, e.g. "€ 80M–120M". */
  annualRevenue?: string
  established?: number
  capabilities?: string[]
  certifications?: string[]
  /** Source provenance: database/local suppliers vs web research. */
  source?: string
  repurchasePriority?: string
  /** URLs fetched during deep supplier research. */
  sourceUrls?: string[]
  /** Short text snippets supporting the extracted supplier profile. */
  evidenceSnippets?: string[]
  /** Product evidence extracted from supplier pages or local purchase records. */
  productName?: string
  brand?: string
  model?: string
  specifications?: string
  standards?: string[]
  /** Commercial details are optional because many web listings omit them. */
  unitPrice?: number | null
  currency?: string
  quoteConditions?: string
  deliveryLeadTime?: string
  paymentTerms?: string
  /** Per-criterion evidence scores returned by the sourcing agent. */
  criteriaScores?: Record<string, number>
  weightedCriteriaScore?: number
  appliedCriteria?: EvaluationCriterion[]
  /** Whether the profile has a contact method and product evidence to verify. */
  verificationStatus?: 'verified' | 'partial' | 'unverified' | string
  dataCompleteness?: number
  /** 0–100 relevance to the query. */
  matchScore: number
}

/** A user-defined sourcing decision dimension and its relative importance. */
export interface EvaluationCriterion {
  key: string
  label: string
  weight: number
}

/**
 * A reusable local supplier record. Unlike a search result, it does not need a
 * relevance score and can be curated by the procurement team over time.
 */
export interface SupplierDirectoryEntry {
  id: string
  name: string
  country?: string
  city?: string
  address?: string
  contactPerson?: string
  phone?: string
  email?: string
  website?: string
  capabilities?: string[]
  certifications?: string[]
  tags: string[]
  notes?: string
  /** Backend provenance of the record; manually curated entries are internal. */
  origin?: 'internal' | 'web' | string
  source?: 'local' | 'database' | string
  preferred?: boolean
  historicalPerformance?: string | number | boolean
  minimumOrderQuantity?: string
  productionCapacity?: string
  environmentalStandards?: string[]
  criteriaScores?: Record<string, number>
  productName?: string
  brand?: string
  model?: string
  specifications?: string
  standards?: string[]
  unitPriceEur?: number | null
  quoteConditions?: string
  deliveryDays?: number | null
  deliveryLabel?: string
  paymentTerm?: string
  paymentLabel?: string
  createdAt?: number
  updatedAt: number
}

export type SupplierDirectoryInput = Omit<
  SupplierDirectoryEntry,
  'id' | 'createdAt' | 'updatedAt'
>

// --- Comparison (standard-product price benchmarking) ----------------------

export type PaymentTermKey = 'onAccount' | 'prepayment' | 'card' | 'unknown'

export interface ComparisonItem {
  id: string
  vendor: string
  platform: string
  product: string
  matchScore: number
  /** Numeric unit price in EUR — used for sorting. Null means web candidate needs manual price check. */
  unitPriceEur: number | null
  /** Original localized price label, e.g. "€ 2.180 / Stk." */
  unitLabel: string
  /** Numeric lead time in days — used for sorting. Null means web candidate needs manual delivery check. */
  deliveryDays: number | null
  deliveryLabel: string
  paymentTerm: PaymentTermKey
  paymentLabel: string
  deliveryMethod: string
  rating: number
  reviews: number
  source?: 'database' | 'web' | string
  sourceDetail?: string
  sourceUrls?: string[]
  evidenceSnippets?: string[]
  priceConfidence?: 'extracted' | 'unknown' | string
  /** Per-criterion evidence scores returned when custom comparison criteria are used. */
  criteriaScores?: Record<string, number>
  weightedCriteriaScore?: number
  appliedCriteria?: EvaluationCriterion[]
}

export type DeliveryOptionKey = 'unlimited' | 'within3' | 'within7'

/** User-set ranking weights (percentages that always sum to 100). */
export interface FactorWeights {
  price: number
  delivery: number
  rating: number
}

// --- Memory + feedback -----------------------------------------------------

export interface FeedbackRecord {
  /** Name of the chosen supplier / vendor. */
  chosenName: string
  /** 0–5 stars each. */
  quality: number
  logistics: number
  priceSatisfaction: number
  service: number
  comment: string
  submittedAt: number
}

/** Structured fields for the sourcing module (双模输入的表单部分). */
export interface StructuredFields {
  productName?: string
  quantity?: string
  unit?: string
  brand?: string
  category?: string
  country?: string
  certifications?: string
}

/** Raw input values needed to re-run / reopen a past conversation. */
export interface ConversationRestore {
  query: string
  minPrice?: string
  maxPrice?: string
  deliveryTime?: DeliveryOptionKey
  /** Explicit target market for a standard-product comparison. */
  comparisonCountry?: string
  weights?: FactorWeights
  /** Structured fields for sourcing module reopen. */
  productName?: string
  quantity?: string
  unit?: string
  brand?: string
  model?: string
  specifications?: string
  standards?: string
  structuredCategory?: string
  structuredCountry?: string
  structuredCerts?: string
  sourcingCriteria?: EvaluationCriterion[]
  comparisonCriteria?: EvaluationCriterion[]
}

/** One logged query + everything the user typed for it. */
export interface ConversationRecord {
  id: string
  module: Extract<ModuleId, 'sourcing' | 'comparison'>
  query: string
  /** Human-readable filter summary for display in the memory list. */
  filters: Record<string, string>
  /** Raw input values so the conversation can be reopened in its module. */
  restore?: ConversationRestore
  /** Snapshot of the full request payload sent to the backend. */
  requestSnapshot?: Record<string, unknown>
  /** Full search results for auto-restore on reopen (not truncated). */
  resultsSnapshot?: Record<string, unknown>[]
  resultCount: number
  /** Names of the candidates returned, so feedback can reference them. */
  candidateNames: string[]
  timestamp: number
  feedback?: FeedbackRecord
}